requirements.txt
black==23.3.0
openai==0.27.7
pre-commit==3.3.2
python-dotenv==1.0.0